♻️ Improve handling of LaunchType, simplify dependency graph
