+++ Sources/AppcuesKit/Utilities/LaunchType.swift
@@ -1,0 +1,15 @@
+//
+//  LaunchType.swift
+//  AppcuesKit
+//
+//  Created by James Ellis on 10/27/21.
+//  Copyright © 2021 Appcues. All rights reserved.
+//
+
+import Foundation
+
+internal enum LaunchType {
+    case install
+    case open
+    case update
+}
